Get rid of static fields, fix tests
